+++ packages/core/src/types/IVolumeViewport.ts
@@ -13,7 +13,10 @@ import { VolumeViewportProperties } from '.';
 export default interface IVolumeViewport extends IViewport {
   useCPURendering: boolean;
   getFrameOfReferenceUID: () => string;
-  getProperties: () => any;
+  /**
+   * Retrieve the viewport properties
+   */
+  getProperties: () => VolumeViewportProperties;
   /**
    * canvasToWorld Returns the world coordinates of the given `canvasPos`
    * projected onto the plane defined by the `Viewport`'s `vtkCamera`'s focal point