+++ packages/core/src/utilities/index.ts
@@ -1,4 +1,7 @@
 import csUtils from './invertRgbTransferFunction';
+import createSigmoidRGBTransferFunction from './createSigmoidRGBTransferFunction';
+import getVoiFromSigmoidRGBTransferFunction from './getVoiFromSigmoidRGBTransferFunction';
+import createLinearRGBTransferFunction from './createLinearRGBTransferFunction';
 import scaleRgbTransferFunction from './scaleRgbTransferFunction';
 import triggerEvent from './triggerEvent';
 import uuidv4 from './uuidv4';
@@ -40,6 +43,9 @@ import * as windowLevel from './windowLevel';
 
 export {
   csUtils as invertRgbTransferFunction,
+  createSigmoidRGBTransferFunction,
+  getVoiFromSigmoidRGBTransferFunction,
+  createLinearRGBTransferFunction,
   scaleRgbTransferFunction,
   triggerEvent,
   imageIdToURI,